Fix reviewers comments

+++ src/main/java/com/example/rest/Employee.java
@@ -3,10 +3,12 @@ package com.example.rest;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class Employee {
     private String name;
     private Long id;
-    private Double salary;
+    private BigDecimal salary;
 }
